requirements.txt
pyspark~=3.0.3
pyspark-stubs~=3.0.0.post3
requests~=2.32.3
fastapi~=0.116.1
confluent-kafka~=2.5.3
python-dotenv~=1.1.1